+++ frontend/src/hooks/useOnContainerScroll.ts
@@ -1,43 +1,10 @@
-import { RefObject, useEffect, useRef, useState } from 'react';
+import useIntersectionObserver from './useIntersectionObserver';
+import { RefObject, useEffect } from 'react';
 
 interface ScrollInfo {
   scrollY: number;
   progress: number;
 }
-
-const useIntersectionObserver = <T extends Element>(targetRef: RefObject<T>, threshold: number = 0) => {
-  const observerRef = useRef<IntersectionObserver>();
-  const [isLoaded, setIsLoaded] = useState(false);
-
-  const intersectionCallBack = (entries: IntersectionObserverEntry[], io: IntersectionObserver) => {
-    entries.forEach(entry => {
-      if (entry.isIntersecting) {
-        setIsLoaded(true);
-        return;
-      }
-      if (!entry.isIntersecting) {
-        setIsLoaded(false);
-        return;
-      }
-    });
-  };
-
-  useEffect(() => {
-    if (!observerRef.current) {
-      observerRef.current = new IntersectionObserver(intersectionCallBack, {
-        threshold,
-      });
-    }
-
-    targetRef.current && observerRef.current.observe(targetRef.current);
-
-    return () => {
-      observerRef.current?.disconnect();
-    };
-  }, [targetRef.current]);
-
-  return { isLoaded };
-};
 
 const useOnContainerScroll = (container: RefObject<HTMLElement>, onScroll: (info: ScrollInfo) => void) => {
   const { isLoaded } = useIntersectionObserver<HTMLElement>(container);
